app/requirements.txt
aiohttp
fake-useragent
nest-asyncio
hydra-core
# psycopg2
newspaper3k
feedparser
peewee
elasticsearch-dsl>=7.0.0,<8.0.0
pytest
requests
requests-cache
get-all-tickers
redis
msgpack
news-please
grpcio
grpcio-tools
